rss update

+++ rss/rss.sh
@@ -9,25 +9,34 @@ folder="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
 
 mkdir -p "${folder}"/tmp
 
-# if output.xml exists, remove it
-if [ -f output.xml ]; then
-    rm output.xml
+# estrai soltanto i bollettini con livello diverso da 0
+mlrgo -S --csv filter -x '$livello=~"0"' "${folder}"/../data/ondate-calore_latest.csv > "${folder}"/tmp/tmp.csv
+
+# se non ci sono bollettini con livello diverso da 0, esci
+if [ ! -s "${folder}"/tmp/tmp.csv ]; then
+    echo "No data"
+    exit 0
 fi
 
-mlrgo -S --csv filter '$livello=~"."' "${folder}"/../data/ondate-calore_latest.csv > "${folder}"/tmp/tmp.csv
-
+# estrai dati anagrafici delle città, utili per il join
 mlrgo -S --csv --from "${folder}"/../data/citta-anagrafica.csv cut -f citta,admin3code,name > "${folder}"/tmp/citta-anagrafica.csv
 
+# aggiungi ai dati dei bollettini i dati anagrafici estratti
 mlrgo -S --csv join --ul -j citta -f "${folder}"/tmp/tmp.csv then unsparsify "${folder}"/tmp/citta-anagrafica.csv > "${folder}"/tmp/rss.csv
 
+# aggiungi ai dati dei bollettini le infi sui livelli di rischio
 mlrgo --csv -S join --ul -j livello -f "${folder}"/tmp/rss.csv then unsparsify "${folder}"/../data/livelli.csv > "${folder}"/tmp/tmp.csv
 
+# rinomina il file
 mv "${folder}"/tmp/tmp.csv "${folder}"/tmp/rss.csv
 
+# crea campi e/o rinominali in modo che ci siano le colonne utili per il feed RSS
 mlrgo -I -S --csv --from "${folder}"/tmp/rss.csv put '$data_title=strftime(strptime($data, "%Y-%m-%d"),"%d/%m/%Y");$link=$URL."?data=".$data;$title=$citta." - ".$data_title.": ".$livello_label;$description=$livello_des;$pubDate=$data." 00:00:00";$category=$livello_label' then sort -r data -f citta then cut -f link,title,description,pubDate,category,admin3code,name
 
-mlrgo --icsv --ojsonl cat "${folder}"/tmp/rss.csv > "${folder}"/tmp/rss.jsonl
+# crea una copia in jsonl
+# mlrgo --icsv --ojsonl cat "${folder}"/tmp/rss.csv > "${folder}"/tmp/rss.jsonl
 
+# per ogni codice città con dei dati, crea un feed RSS
 mlrgo --c2n cut -f admin3code then uniq -a "${folder}"/tmp/rss.csv | while read -r admin3code; do
   echo "admin3code: $admin3code"
   mkdir -p "${folder}"/../docs/rss
